Trait cp_permastore::PermastoreApi

pub trait PermastoreApi<Block: BlockT, BlockNumber, ExtrinsicIndex, Hash>: Core<Block> where
    BlockNumber: Codec,
    ExtrinsicIndex: Codec,
    Hash: Codec, {
    fn chunk_root(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        block_number: BlockNumber, 
        extrinsic_index: ExtrinsicIndex
    ) -> Result<Option<Hash>, ApiError> { ... }

    fn chunk_root_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        block_number: BlockNumber, 
        extrinsic_index: ExtrinsicIndex
    ) -> Result<Option<Hash>, ApiError> { ... }

    fn find_recall_block(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        recall_byte: u64
    ) -> Result<Option<BlockNumber>, ApiError> { ... }

    fn find_recall_block_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        recall_byte: u64
    ) -> Result<Option<BlockNumber>, ApiError> { ... }

    fn data_size(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        block_number: BlockNumber, 
        extrinsic_index: ExtrinsicIndex
    ) -> Result<u32, ApiError> { ... }

    fn data_size_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        block_number: BlockNumber, 
        extrinsic_index: ExtrinsicIndex
    ) -> Result<u32, ApiError> { ... }

    fn require_proof_of_access(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<bool, ApiError> { ... }

    fn require_proof_of_access_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<bool, ApiError> { ... }

    fn block_size(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, ApiError> { ... }

    fn block_size_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, ApiError> { ... }

    fn weave_size(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, ApiError> { ... }

    fn weave_size_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, ApiError> { ... }
}

The permastore API.

Provided methods

fn chunk_root(
&self,
__runtime_api_at_param__: &BlockId<Block>,
block_number: BlockNumber,
extrinsic_index: ExtrinsicIndex
) -> Result<Option<Hash>, ApiError>

Returns the chunk root given block_number and extrinsic_index.

fn chunk_root_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
block_number: BlockNumber,
extrinsic_index: ExtrinsicIndex
) -> Result<Option<Hash>, ApiError>

Returns the chunk root given block_number and extrinsic_index.

fn find_recall_block(
&self,
__runtime_api_at_param__: &BlockId<Block>,
recall_byte: u64
) -> Result<Option<BlockNumber>, ApiError>

Returns the number of block in which the recall byte is included.

fn find_recall_block_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
recall_byte: u64
) -> Result<Option<BlockNumber>, ApiError>

Returns the number of block in which the recall byte is included.

fn data_size(
&self,
__runtime_api_at_param__: &BlockId<Block>,
block_number: BlockNumber,
extrinsic_index: ExtrinsicIndex
) -> Result<u32, ApiError>

Returns the size of transaction data given block_number and extrinsic_index.

fn data_size_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
block_number: BlockNumber,
extrinsic_index: ExtrinsicIndex
) -> Result<u32, ApiError>

Returns the size of transaction data given block_number and extrinsic_index.

fn require_proof_of_access(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<bool, ApiError>

Returns true if the proof of access is required for the block.

fn require_proof_of_access_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<bool, ApiError>

Returns true if the proof of access is required for the block.

fn block_size(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, ApiError>

Returns the size of current block.

fn block_size_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, ApiError>

Returns the size of current block.

fn weave_size(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, ApiError>

Returns the size of entire weave.

fn weave_size_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, ApiError>

Returns the size of entire weave.